+++ tests/test_routers/test_images.py
@@ -7,47 +7,49 @@ from tests.fixtures import (get_admin_header, get_user_header,  # noqa: F401
 """TEST POST REQUESTS"""
 
 
-def test_post_image(set_test_environment):  # noqa: F811
+def test_post_image_by_admin(set_test_environment, get_admin_header):  # noqa: F811
     with open('./tests/media/image.jpg', 'rb') as file:
         image_file = file.read()
     response = client.post(
         '/images/?product_id=1',
         files={'file': ('test.jpg', image_file, 'image/jpeg')},
-        # headers={'Authorization': get_admin_header}
+        headers={'Authorization': get_admin_header}
     )
     assert response.status_code == 201
 
 
-def test_post_incorrect_image(set_test_environment):  # noqa: F811
+def test_post_incorrect_image_by_admin(set_test_environment, get_admin_header):  # noqa: F811
     with open('./tests/media/incorrect_image.jpg', 'rb') as file:
         image_file = file.read()
     response = client.post(
         '/images/?product_id=1',
         files={'file': ('test.jpg', image_file, 'image/jpeg')},
-        # headers={'Authorization': get_admin_header}
+        headers={'Authorization': get_admin_header}
     )
     assert response.status_code == 422
 
 
-def test_post_image_with_incorrect_product_id(set_test_environment):  # noqa: F811
+def test_post_image_with_incorrect_product_id(set_test_environment, get_admin_header):  # noqa: F811
     with open('./tests/media/image.jpg', 'rb') as file:
         image_file = file.read()
     response = client.post(
         '/images/?product_id=5',
         files={'file': ('test.jpg', image_file, 'image/jpeg')},
-        # headers={'Authorization': get_admin_header}
+        headers={'Authorization': get_admin_header}
     )
     assert response.status_code == 422
 
-# def test_post_image_by_user(set_test_environment):  # noqa: F811
-#     with open('./tests/media/image.jpg', 'rb') as file:
-#         image_file = file.read()
-#     response = client.post(
-#         '/images/',
-#         data={'product_id': 1},
-#         files={'file': ('test.jpg', image_file, 'image/jpeg')},
-#     )
-#     assert response.status_code == 401
+
+def test_post_image_by_user(set_test_environment, get_user_header):  # noqa: F811
+    with open('./tests/media/image.jpg', 'rb') as file:
+        image_file = file.read()
+    response = client.post(
+        '/images/',
+        data={'product_id': 1},
+        files={'file': ('test.jpg', image_file, 'image/jpeg')},
+        headers={'Authorization': get_user_header}
+    )
+    assert response.status_code == 401
 
 
 """TEST GET REQUESTS"""
@@ -99,15 +101,15 @@ def test_get_nonexistent_image(set_test_environment):  # noqa: F811
 """TEST DELETE REQUESTS"""
 
 
-# def test_delete_image_by_user(get_user_header, set_test_environment):  # noqa: F811
-#     response = client.delete('/images/1/', headers={'Authorization': get_user_header})
-#     assert response.status_code == 401
+def test_delete_image_by_user(set_test_environment, get_user_header):  # noqa: F811
+    response = client.delete('/images/1/', headers={'Authorization': get_user_header})
+    assert response.status_code == 401
 
 
-def test_delete_image(set_test_environment):  # noqa: F811
+def test_delete_image_by_admin(set_test_environment, get_admin_header):  # noqa: F811
     response = client.get('/images/1/')
     url = response.json()['url']
-    response = client.delete('/images/1/', headers={})
+    response = client.delete('/images/1/', headers={'Authorization': get_admin_header})
 
     assert response.status_code == 202
     assert response.json() is None
@@ -120,8 +122,8 @@ def test_delete_image(set_test_environment):  # noqa: F811
     assert not path.exists(f'./tests/media/{url}')
 
 
-def test_delete_image_with_incorrect_id(set_test_environment):  # noqa: F811
-    response = client.delete('/images/1/', headers={})
+def test_delete_image_with_incorrect_id(set_test_environment, get_admin_header):  # noqa: F811
+    response = client.delete('/images/1/', headers={'Authorization': get_admin_header})
 
     assert response.status_code == 202
     assert response.json() is None
