+++ tests/test_routers/test_products.py
@@ -1,5 +1,5 @@
 from tests.client import client
-from tests.fixtures import set_test_environment # noqa: F401
+from tests.fixtures import get_admin_header, get_user_header, set_test_environment  # noqa: F401
 
 """TEST GET REQUESTS"""
 
@@ -102,7 +102,7 @@ def test_get_product(set_test_environment):  # noqa: F811
 """TEST POST REQUESTS"""
 
 
-def test_post_product(set_test_environment):  # noqa: F811
+def test_post_product(set_test_environment, get_admin_header):  # noqa: F811
     response = client.post(
         '/products/',
         json={'name': 'Product 2',
@@ -110,7 +110,7 @@ def test_post_product(set_test_environment):  # noqa: F811
               'constitution': 'Constitution of Product 2',
               'price': 99.99,
               'category_id': 2},
-        # headers={'Authorization': get_admin_header}
+        headers={'Authorization': get_admin_header}
     )
 
     assert response.status_code == 201
@@ -128,38 +128,38 @@ def test_post_product(set_test_environment):  # noqa: F811
                                'images': []}
 
 
-def test_post_product_with_invalid_data(set_test_environment):  # noqa: F811
+def test_post_product_with_invalid_data(set_test_environment, get_admin_header):  # noqa: F811
     response = client.post(
         '/products/',
         json={'username': 213123},
-        # headers={'Authorization': get_admin_header}
+        headers={'Authorization': get_admin_header}
     )
 
     assert response.status_code == 422
 
 
-# def test_post_product_by_user(set_test_environment):  # noqa: F811
-#     response = client.post(
-#         '/products/',
-#         json={'name': 'Product 2',
-#               'description': 'Description of Product 2',
-#               'constitution': 'Constitution of Product 2',
-#               'price': 199.99,
-#               'category_id': 2},
-#         # headers={'Authorization': get_user_header}
-#     )
-#
-#     assert response.status_code == 401
+def test_post_product_by_user(set_test_environment, get_user_header):  # noqa: F811
+    response = client.post(
+        '/products/',
+        json={'name': 'Product 2',
+              'description': 'Description of Product 2',
+              'constitution': 'Constitution of Product 2',
+              'price': 199.99,
+              'category_id': 2},
+        headers={'Authorization': get_user_header}
+    )
+
+    assert response.status_code == 401
 
 
 """TEST PUT REQUESTS"""
 
 
-def test_put_product_by_admin(set_test_environment):  # noqa: F811
+def test_put_product_by_admin(set_test_environment, get_admin_header):  # noqa: F811
     response = client.put(
         '/products/2/',
         json={'name': 'Product 2 updated'},
-        # headers={'Authorization': get_admin_header}
+        headers={'Authorization': get_admin_header}
     )
 
     assert response.status_code == 202
@@ -177,20 +177,20 @@ def test_put_product_by_admin(set_test_environment):  # noqa: F811
                                'images': []}
 
 
-# def test_put_product_by_user(set_test_environment):  # noqa: F811
-#     response = client.put(
-#         '/products/2/',
-#         json={'name': 'Product 2 updated'},
-#         # headers={'Authorization': get_user_header}
-#     )
-#     assert response.status_code == 401
-#
-#
-def test_put_product_with_invalid_data(set_test_environment):  # noqa: F811
+def test_put_product_by_user(set_test_environment, get_user_header):  # noqa: F811
+    response = client.put(
+        '/products/2/',
+        json={'name': 'Product 2 updated'},
+        headers={'Authorization': get_user_header}
+    )
+    assert response.status_code == 401
+
+
+def test_put_product_with_invalid_data(set_test_environment, get_admin_header):  # noqa: F811
     response = client.put(
         '/products/2/',
         json={'number': '88000965599'},
-        # headers={'Authorization': get_admin_header}
+        headers={'Authorization': get_admin_header}
     )
     assert response.status_code == 422
 
@@ -198,8 +198,8 @@ def test_put_product_with_invalid_data(set_test_environment):  # noqa: F811
 """TEST DELETE REQUESTS"""
 
 
-def test_delete_product(set_test_environment):  # noqa: F811
-    response = client.delete('/products/2/')
+def test_delete_product(set_test_environment, get_admin_header):  # noqa: F811
+    response = client.delete('/products/2/', headers={'Authorization': get_admin_header})
 
     assert response.status_code == 202
     assert response.json() is None
@@ -210,13 +210,13 @@ def test_delete_product(set_test_environment):  # noqa: F811
     assert response.json() is None
 
 
-# def test_delete_product_by_user(set_test_environment):  # noqa: F811
-#     response = client.delete('/products/1/')
-#     assert response.status_code == 401
-#
-#
-# def test_delete_product_by_with_invalid_id(set_test_environment):  # noqa: F811
-#     response = client.delete('/products/3/')
-#
-#     assert response.status_code == 202
-#     assert response.json() is None
+def test_delete_product_by_user(set_test_environment, get_user_header):  # noqa: F811
+    response = client.delete('/products/1/', headers={'Authorization': get_user_header})
+    assert response.status_code == 401
+
+
+def test_delete_product_with_invalid_id(set_test_environment, get_admin_header):  # noqa: F811
+    response = client.delete('/products/3/', headers={'Authorization': get_admin_header})
+
+    assert response.status_code == 202
+    assert response.json() is None
