+++ tests/test_routers/test_users.py
@@ -5,8 +5,8 @@ from tests.fixtures import (get_admin_header, get_user_header,  # noqa: F401
 """TEST GET REQUESTS"""
 
 
-def test_get_users_by_admin(set_test_environment):  # noqa: F811
-    response = client.get('/users/', headers={})
+def test_get_users_by_admin(set_test_environment, get_admin_header):  # noqa: F811
+    response = client.get('/users/', headers={'Authorization': get_admin_header})
     assert response.status_code == 200
 
     response = response.json()
@@ -28,8 +28,8 @@ def test_get_users_by_admin(set_test_environment):  # noqa: F811
     ]
 
 
-def test_get_admin_by_admin(set_test_environment):  # noqa: F811
-    response = client.get('/users/1/', headers={})
+def test_get_admin_by_admin(set_test_environment, get_admin_header):  # noqa: F811
+    response = client.get('/users/1/', headers={'Authorization': get_admin_header})
     assert response.status_code == 200
 
     response = response.json()
@@ -43,8 +43,8 @@ def test_get_admin_by_admin(set_test_environment):  # noqa: F811
     }
 
 
-def test_get_user_by_admin(set_test_environment):  # noqa: F811
-    response = client.get('/users/2/', headers={})
+def test_get_user_by_admin(set_test_environment, get_admin_header):  # noqa: F811
+    response = client.get('/users/2/', headers={'Authorization': get_admin_header})
     assert response.status_code == 200
 
     response = response.json()
@@ -58,33 +58,33 @@ def test_get_user_by_admin(set_test_environment):  # noqa: F811
     }
 
 
-# def test_get_users_by_user(get_user_header, set_test_environment):  # noqa: F811
-#     response = client.get('/users/', headers={'Authorization': get_user_header})
-#     assert response.status_code == 401
-#
-#
-# def test_get_user_by_user(get_user_header, set_test_environment):  # noqa: F811
-#     response = client.get('/users/2/', headers={'Authorization': get_user_header})
-#     assert response.status_code == 200
-#
-#     response = response.json()
-#     response.pop('date_registration')
-#
-#     assert response == {
-#         'id': 2,
-#         'username': 'user',
-#         'number': '99999999999',
-#         'is_admin': False
-#     }
+def test_get_users_by_user(set_test_environment, get_user_header):  # noqa: F811
+    response = client.get('/users/', headers={'Authorization': get_user_header})
+    assert response.status_code == 401
 
 
-# def test_get_another_user_by_user(get_user_header, set_test_environment):  # noqa: F811
-#     response = client.get('/users/1/', headers={'Authorization': get_user_header})
-#     assert response.status_code == 401
+def test_get_user_by_user(set_test_environment, get_user_header):  # noqa: F811
+    response = client.get('/users/2/', headers={'Authorization': get_user_header})
+    assert response.status_code == 200
+
+    response = response.json()
+    response.pop('date_registration')
+
+    assert response == {
+        'id': 2,
+        'username': 'user',
+        'number': '99999999999',
+        'is_admin': False
+    }
 
 
-def test_get_nonexistent_user(set_test_environment):  # noqa: F811
-    response = client.get('/users/3/', headers={})
+def test_get_another_user_by_user(set_test_environment, get_user_header):  # noqa: F811
+    response = client.get('/users/1/', headers={'Authorization': get_user_header})
+    assert response.status_code == 401
+
+
+def test_get_nonexistent_user(set_test_environment, get_admin_header):  # noqa: F811
+    response = client.get('/users/3/', headers={'Authorization': get_admin_header})
     assert response.status_code == 200
     assert response.json() is None
 
@@ -92,7 +92,7 @@ def test_get_nonexistent_user(set_test_environment):  # noqa: F811
 """TEST POST REQUESTS"""
 
 
-def test_post_user(set_test_environment):  # noqa: F811
+def test_post_user(set_test_environment, get_admin_header):  # noqa: F811
     response = client.post(
         '/users/registration/',
         json={
@@ -105,7 +105,7 @@ def test_post_user(set_test_environment):  # noqa: F811
     assert response.status_code == 201
     assert response.json() is None
 
-    response = client.get("/users/3/", headers={})
+    response = client.get("/users/3/", headers={'Authorization': get_admin_header})
     assert response.status_code == 200
 
     response = response.json()
@@ -119,7 +119,7 @@ def test_post_user(set_test_environment):  # noqa: F811
     }
 
 
-def test_post_user_with_invalid_data(set_test_environment):  # noqa: F811
+def test_post_user_with_invalid_data(set_test_environment, get_admin_header):  # noqa: F811
     response = client.post(
         '/users/registration/',
         json={
@@ -127,7 +127,8 @@ def test_post_user_with_invalid_data(set_test_environment):  # noqa: F811
             'password': 'password',
             'password2': 'password',
             'number': '88888888888'
-        }
+        },
+        headers={'Authorization': get_admin_header}
     )
     assert response.status_code == 422
 
@@ -135,16 +136,16 @@ def test_post_user_with_invalid_data(set_test_environment):  # noqa: F811
 """TEST PUT REQUESTS"""
 
 
-def test_put_user_by_admin(set_test_environment):  # noqa: F811
+def test_put_user_by_admin(set_test_environment, get_admin_header):  # noqa: F811
     response = client.put(
         '/users/3/',
         json={'username': 'user_3_updated'},
-        headers={}
+        headers={'Authorization': get_admin_header},
     )
     assert response.status_code == 202
     assert response.json() is None
 
-    response = client.get("/users/3/", headers={})
+    response = client.get("/users/3/", headers={'Authorization': get_admin_header})
     assert response.status_code == 200
 
     response = response.json()
@@ -158,43 +159,43 @@ def test_put_user_by_admin(set_test_environment):  # noqa: F811
     }
 
 
-# def test_put_user_by_another_user(get_user_header, set_test_environment):  # noqa: F811
-#     response = client.put(
-#         '/users/3/',
-#         json={'username': 'user_3_updated_again'},
-#         headers={'Authorization': get_user_header}
-#     )
-#     assert response.status_code == 401
+def test_put_user_by_another_user(set_test_environment, get_user_header):  # noqa: F811
+    response = client.put(
+        '/users/3/',
+        json={'username': 'user_3_updated_again'},
+        headers={'Authorization': get_user_header}
+    )
+    assert response.status_code == 401
 
 
-# def test_put_user_by_user(get_user_header, set_test_environment):  # noqa: F811
-#     response = client.put(
-#         '/users/2/',
-#         json={'username': 'user_2_updated'},
-#         headers={'Authorization': get_user_header}
-#     )
-#     assert response.status_code == 202
-#     assert response.json() is None
-#
-#     response = client.get('/users/2/', headers={'Authorization': get_user_header})
-#     assert response.status_code == 200
-#
-#     response = response.json()
-#     response.pop('date_registration')
-#
-#     assert response == {
-#         'id': 2,
-#         'username': 'user_2_updated',
-#         'number': '99999999999',
-#         'is_admin': False
-#     }
+def test_put_user_by_user(set_test_environment, get_user_header):  # noqa: F811
+    response = client.put(
+        '/users/2/',
+        json={'username': 'user_2_updated'},
+        headers={'Authorization': get_user_header}
+    )
+    assert response.status_code == 202
+    assert response.json() is None
+
+    response = client.get('/users/2/', headers={'Authorization': get_user_header})
+    assert response.status_code == 200
+
+    response = response.json()
+    response.pop('date_registration')
+
+    assert response == {
+        'id': 2,
+        'username': 'user_2_updated',
+        'number': '99999999999',
+        'is_admin': False
+    }
 
 
-def test_put_user_with_invalid_data(set_test_environment):  # noqa: F811
+def test_put_user_with_invalid_data(set_test_environment, get_admin_header):  # noqa: F811
     response = client.put(
         '/users/3/',
         json={'number': '99999999999'},
-        headers={}
+        headers={'Authorization': get_admin_header},
     )
     assert response.status_code == 422
 
@@ -202,28 +203,28 @@ def test_put_user_with_invalid_data(set_test_environment):  # noqa: F811
 """TEST DELETE REQUESTS"""
 
 
-def test_delete_user_by_admin(set_test_environment):  # noqa: F811
-    response = client.delete('/users/3/', headers={})
+def test_delete_user_by_admin(set_test_environment, get_admin_header):  # noqa: F811
+    response = client.delete('/users/3/', headers={'Authorization': get_admin_header})
     assert response.status_code == 202
     assert response.json() is None
 
-    response = client.get('/users/3/', headers={})
+    response = client.get('/users/3/', headers={'Authorization': get_admin_header})
     assert response.status_code == 200
 
     assert response.json() is None
 
 
-# def test_delete_user_by_user(set_test_environment):  # noqa: F811
-#     response = client.delete('/users/2/', headers={})
-#     assert response.status_code == 202
-#     assert response.json() is None
-#
-#     response = client.get('/users/2/', headers={})
-#     assert response.status_code == 200
-#
-#     assert response.json() is None
+def test_delete_user_by_user(set_test_environment, get_user_header, get_admin_header):  # noqa: F811
+    response = client.delete('/users/2/', headers={'Authorization': get_user_header})
+    assert response.status_code == 202
+    assert response.json() is None
+
+    response = client.get('/users/2/', headers={'Authorization': get_admin_header})
+    assert response.status_code == 200
+
+    assert response.json() is None
 
 
-# def test_delete_user_by_another_user(get_user_header, set_test_environment):  # noqa: F811
-#     response = client.delete('/users/1/', headers={'Authorization': get_user_header})
-#     assert response.status_code == 401
+def test_delete_user_by_another_user(set_test_environment, get_user_header):  # noqa: F811
+    response = client.delete('/users/1/', headers={'Authorization': get_user_header})
+    assert response.status_code == 401
